Move Bond class in archiveinfo module
