+++ py7zr/archiveinfo.py
@@ -35,7 +35,7 @@ from typing import Any, BinaryIO, Dict, List, Optional, Tuple
 
 from py7zr.compressor import SevenZipCompressor, SevenZipDecompressor
 from py7zr.exceptions import Bad7zFile, InternalError
-from py7zr.helpers import ArchiveTimestamp, Bond, calculate_crc32
+from py7zr.helpers import ArchiveTimestamp, calculate_crc32
 from py7zr.properties import ENCODED_HEADER_DEFAULT, ENCRYPTED_HEADER_DEFAULT, MAGIC_7Z, CompressionMethod, Property
 
 MAX_LENGTH = 65536
@@ -284,6 +284,18 @@ class PackInfo:
             for crc in self.crcs:
                 write_uint64(file, crc)
         write_byte(file, Property.END)
+
+
+class Bond:
+    """Represent bindings between two methods.
+    bonds[i] = (incoder, outstream)
+    means
+    methods[i].stream[outstream] output data go to method[incoder].stream[0]
+    """
+
+    def __init__(self, incoder, outcoder):
+        self.incoder = incoder
+        self.outcoder = outcoder
 
 
 class Folder:
